Moved conversions of EVerest to/from libevse-security types to lib/staging in order to be able to use it also within other modules
