+++ lib/staging/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectory(can_dpm1000)
+add_subdirectory(evse_security)
 if(EVEREST_DEPENDENCY_ENABLED_LIBSLAC AND EVEREST_DEPENDENCY_ENABLED_LIBFSM)
     add_subdirectory(slac)
 endif()
