+++ modules/EvseSecurity/CMakeLists.txt
@@ -13,13 +13,13 @@ ev_setup_cpp_module()
 target_link_libraries(${MODULE_NAME}
     PRIVATE
         everest::evse_security
+        everest::evse_security_conversions
 )
 # ev@bcc62523-e22b-41d7-ba2f-825b493a3c97:v1
 
 target_sources(${MODULE_NAME}
     PRIVATE
         "main/evse_securityImpl.cpp"
-        "conversions.cpp"
 )
 
 # ev@c55432ab-152c-45a9-9d2e-7281d50c69c3:v1
